+++ packages/stages/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./types.js";
 export { STUB_VOICES, type VoiceDescriptor } from "./voice.js";
-export { runScript } from "./script.js";
+export { createScriptRunner, runScript, ScriptGenerationError } from "./script.js";
 export { runVoice } from "./voice.js";
 export { runFootage, type FootageRunInput } from "./footage.js";
 export { runAlign } from "./align.js";